requirements.txt
asgiref==3.6.0
Django==4.1.5
django-cors-headers==3.13.0
django-environ==0.9.0
django-environ-2==2.3.0
djangorestframework==3.14.0
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
